Simplify validator code
